refactor: moves data-source.ts, database.providers.ts and the migrations folder into src/database and fixes the migration:generate and migration:run scripts
