+++ backend/package.json
@@ -15,8 +15,8 @@
     "services:down": "docker compose -f infra/docker/compose.yaml down && npm run services:clear:volumes",
     "services:clear:volumes": "docker volume rm docker_mysql_data docker_workbench_config",
     "services:wait:database": "node infra/scripts/wait-for-mysql.mjs",
-    "migration:generate": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js migration:generate -d src/data-source.ts src/migrations/InitialSchema",
-    "migration:run": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js migration:run -d src/data-source.ts",
+    "migration:generate": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js migration:generate -d src/database/data-source.ts src/database/migrations/InitialSchema",
+    "migration:run": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js migration:run -d src/database/data-source.ts",
     "migration:revert": "npx typeorm migration:revert",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
