+++ backend/src/leads/leads.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { LeadsService } from './leads.service';
 import { LeadsController } from './leads.controller';
 import { DatabaseModule } from 'src/database/database.module';
-import { leadProviders } from 'src/database.providers';
+import { leadProviders } from 'src/database/database.providers';
 import { PdfModule } from 'src/pdf/pdf.module';
 
 @Module({
